new funct mpi...

+++ fembin/mpi_basin.sh
@@ -18,9 +18,18 @@ newnode=$basin.$npart.node
 visual=$basin.$npart.visual
 
 #---------------------------------------------------
+# cleaning files
+#---------------------------------------------------
+
+[ -f $newnode ] && rm -f $newnode
+[ -f npart.grd ] && rm -f npart.grd
+[ -f $basin.bas ] && rm -f $basin.bas
+
+#---------------------------------------------------
 # create partitioning
 #---------------------------------------------------
 
+echo "running shyparts..."
 shyparts  -nparts $npart $basin.grd
 
 if [ $? -ne 0 -o ! -f $newnode.grd ]; then
@@ -32,14 +41,26 @@ fi
 # create grd-file for visualization - use "grid -ofT grd-file"
 #---------------------------------------------------
 
+echo "running shybas..."
 shybas -silent -npart $newnode.grd
+
+if [ ! -f npart.grd ]; then
+  echo "*** error creating file npart.grd"
+  exit 1
+fi
 mv npart.grd $visual.grd
 
 #---------------------------------------------------
 # create bas file
 #---------------------------------------------------
 
+echo "running shypre..."
 shypre -silent -partition $newnode.grd $basin.grd
+
+if [ ! -f $basin.bas ]; then
+  echo "*** error creating file $basin.bas"
+  exit 1
+fi
 
 #---------------------------------------------------
 # final info
